+++ CHANGELOG.md
@@ -1,1 +1,11 @@
 # Change Log
+
+## 1.0.0 Beta 1 (02/01/2017)
+
+### Features
+
+- Automates the sending of Ahead of the Curve achievments when signing up for raid groups via the LFG system.
+- Currently only works for Legion raids.
+- Determines highest achievment found for the raid your signing up for on your account (Cutting Edge, Ahead of the Curve, Normal/LFR Achievement).
+- Allows you to override defaults by picking any achievement you have completed. This override will send for any Legion raid you sign up for.
+- When signing up for a group you have the option to send a whisper or not. If no default achievement is found and no override is set no whisper will be sent.